receipt, audittrail

+++ ng-app/src/app/pages/main/main.component.ts
@@ -1,27 +1,49 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { ConfirmationService, MessageService } from 'primeng/api';
+import { BackendService } from 'src/app/service/backend.service';
 import { TokenService } from 'src/app/service/token.service';
 
 @Component({
   selector: 'app-main',
   templateUrl: './main.component.html',
-  styleUrls: ['./main.component.css']
+  styleUrls: ['./main.component.css'],
 })
 export class MainComponent implements OnInit {
-
   constructor(
     private confirmationService: ConfirmationService,
     private messageService: MessageService,
     private router: Router,
     private tokenService: TokenService,
+    private service: BackendService
+  ) {}
 
-  ) { }
+  usermodal: boolean;
+  user: any;
+  position: string;
+  themeicon: string;
+  mode: string;
+  confpassword: string;
 
   ngOnInit(): void {
+    this.confpassword = '';
+    /*this.mode = this.themeService.getCurrentmode();
+    if (this.mode == 'dark') {
+      this.themeicon = 'pi pi-sun';
+    } else {
+      this.themeicon = 'pi pi-moon';
+    }*/
+    this.position = 'bottom';
+    this.usermodal = false;
+    this.user = {};
+    this.getuser();
   }
 
-
+  showprofile() {
+    this.usermodal = true;
+    this.confpassword = '';
+    this.user.password = '';
+  }
 
   onlogout() {
     this.confirmationService.confirm({
@@ -29,10 +51,68 @@ export class MainComponent implements OnInit {
       header: 'Confirmation',
       icon: 'pi pi-exclamation-triangle',
       accept: () => {
-        this.tokenService.destroy();
-        this.router.navigate(['/login']);
-      }
+        this.logout();
+      },
     });
   }
 
+  getuser() {
+    this.service.getusers().subscribe(
+      (res) => {
+        let list: any[] = res;
+        this.user = list.find((i) => {
+          return i.username == this.tokenService.getUser();
+        });
+      },
+      (err) => {
+        this.tokenService.checkSession(err);
+      }
+    );
+  }
+
+  changepassword() {
+    this.confirmationService.confirm({
+      message: 'Update password',
+      header: 'Confirmation',
+      icon: 'pi pi-exclamation-triangle',
+      accept: () => {
+        this.service.changepassword(this.user).subscribe(
+          (res) => {
+            if (res.flag == 'success') {
+              this.messageService.add({
+                key: 'bc',
+                severity: 'success',
+                summary: 'Success',
+                detail: res.event,
+              });
+              this.ngOnInit();
+            }
+          },
+          (err) => {
+            this.messageService.add({
+              key: 'bc',
+              severity: 'error',
+              summary: 'Failed',
+              detail: err.message,
+            });
+            this.tokenService.checkSession(err);
+          }
+        );
+      },
+    });
+  }
+
+  logout() {
+    this.service.userlogout(this.tokenService.getUser()).subscribe(
+      (res) => {
+        if (res.flag == 'success') {
+          this.tokenService.destroy();
+          this.router.navigate(['/login']);
+        }
+      },
+      (err) => {
+        this.tokenService.checkSession(err);
+      }
+    );
+  }
 }
